\name{icossy-package}
\alias{icossy-package}
\docType{package}
\title{Cossy in Internet.}
\description{This package helps web development of cossy.}
\details{
\tabular{ll}{
Package: \tab icossy\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-07-08\cr
License: \tab GPL-2\cr
}
This package creates intermediate interfaces to call cossy functions from web servers. It also formats the output for easy web development.
}
\author{
Ashis Saha

Maintainer: Ashis Saha <ashis@korea.ac.kr>
}

\references{
Saha A, Tan AC, Kang J (2014) Automatic Context-Specific Subnetwork Discovery from Large Interaction Networks. PLoS ONE 9(1): e84227. doi:10.1371/journal.pone.0084227
}
\keyword{ package }
